break cyclic dependency
