+++ protostar/argument_parser/__init__.py
@@ -1,4 +1,4 @@
-from .arg_type import ArgTypeName, StandardParserFactory
+from .arg_type import StandardParserFactory
 from .argument import Argument
 from .argument_parser_facade import (
     ArgumentParserFacade,
@@ -7,3 +7,4 @@ from .argument_parser_facade import (
 from .cli_app import CLIApp
 from .command import Command
 from .config_file_argument_resolver import ConfigFileArgumentResolverProtocol
+from .types import ParserFactoryProtocol, ArgTypeName
